refactor: Rename DataTablePrimary to DataTable
